backend/requirements.txt
# FastAPI & ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pinned for passlib compatibility

# HTTP Client
httpx>=0.26.0

# Vector Database
qdrant-client>=1.7.0,<1.8.0  # Match Qdrant server 1.7.x

# LangChain (for text splitting)
langchain-text-splitters>=0.3.0

# Development
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
